db update in progress

+++ app/src/main/java/ru/pushapptest/todolist/database/models/TodoDB.java
@@ -7,12 +7,12 @@ import androidx.room.PrimaryKey;
 public class TodoDB {
 
     @PrimaryKey
-    public final int id;
+    public final long id;
     public final String headText;
     public final String mainText;
     public final String status;
 
-    public TodoDB(int id, String headText, String mainText, String status) {
+    public TodoDB(long id, String headText, String mainText, String status) {
         this.id = id;
         this.headText = headText;
         this.mainText = mainText;
